[Refactor]#2 Model클래스들을 domain.model 패키지로 이동
