+++ app/src/main/java/com/gallery/kakaogallery/presentation/ui/searchimage/SearchImagesAdapter.kt
@@ -12,7 +12,7 @@ import androidx.recyclerview.widget.RecyclerView
 import com.gallery.kakaogallery.R
 import com.gallery.kakaogallery.databinding.ItemSearchImageBinding
 import com.gallery.kakaogallery.databinding.ItemSearchQueryBinding
-import com.gallery.kakaogallery.model.ImageModel
+import com.gallery.kakaogallery.domain.model.ImageModel
 import com.gallery.kakaogallery.presentation.application.KakaoGalleryApplication
 import com.gallery.kakaogallery.presentation.ui.base.BaseViewHolder
 import com.gallery.kakaogallery.presentation.viewmodel.SearchImageViewModel
