+++ app/src/main/res/layout/item_search_image.xml
@@ -8,7 +8,7 @@
             type="com.gallery.kakaogallery.presentation.viewmodel.SearchImageViewModel" />
         <variable
             name="item"
-            type="com.gallery.kakaogallery.model.ImageModel" />
+            type="com.gallery.kakaogallery.domain.model.ImageModel" />
         <variable
             name="position"
             type="Integer" />
